+++ src/main/kotlin/execution/ProgramCounter.kt
@@ -3,14 +3,14 @@ package dev.apollointhehouse.execution
 import dev.apollointhehouse.asm.Address
 
 class ProgramCounter(val memory: Memory) {
-    private var addr: Int = 0
+    private var addr: Short = 0
 
     fun clock() {
         addr++
     }
 
-    fun load(): Int = memory.load(Address.Raw(addr))
+    fun load(): Short = memory.load(Address.Raw(addr))
     fun set(address: Address.Raw) {
-        addr = address.value - 1
+        addr = (address.value - 1).toShort()
     }
 }
